Move Forms into individual folders
Moves premade forms into individual folders to better control clutter with partial classes.

Fixes namespace issues from form refactor.
