+++ LibraryApp-main/PersonalLibraryApp/PersonalLibraryApplication/Forms/LibraryManager/LibraryManager.resx
